db: move obsolete files code
Move code related to deleting obsolete files from `compaction.go` to
`obsolete_files.go`.
